[BT-undefined] centralize mock data and clean up code architecture
