--- app/components/types/user-card.types.ts
@@ -1,16 +1,0 @@
-export interface Skill {
-  title: string;
-  level: number;
-}
-
-export interface Goal {
-  title: string;
-  level: number;
-}
-
-export interface UserCardProps {
-  name: string;
-  avatarUrl: string;
-  skills: Skill[];
-  currentGoals: Goal[];
-}
