+++ app/features/users/UserPermissionsModal/CustomPermissionsSection.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+// import { useState } from 'react'; // Currently unused
 import { Button } from '@app/components/ui/button';
 import { Badge } from '@app/components/ui/badge';
 import { PenSquare } from 'lucide-react';
